video as class skeleton
