+++ examples/one-scene-input.html
@@ -16,7 +16,6 @@
 
     <script src="../dist/player.js" type="text/javascript"></script>
     <script src="../src/module/scripting.js" type="text/javascript"></script>
-    <script src="../src/module/video.js" type="text/javascript"></script>
     <script src="../src/import/animatron-importer.js" type="text/javascript"></script>
 
     <!-- <script src="../dist/bundle/animatron.js" type="text/javascript"></script> -->
