+++ src/anm/media/video.js
@@ -1,0 +1,27 @@
+/*
+ * Copyright (c) 2011-@COPYRIGHT_YEAR by Animatron.
+ * All rights are reserved.
+ *
+ * Animatron Player is licensed under the MIT License, see LICENSE.
+ *
+ * @VERSION
+ */
+
+  // This module is just a stub template for the moment
+
+var C = require('../constants.js');
+
+C.ET_VIDEO = 'video';
+
+var m_ctx = {};
+
+function Video(src) {
+
+}
+
+Video.prototype.load = function(player) {};
+Video.prototype.apply = function(ctx) {};
+Video.prototype.bounds = function() {};
+Video.prototype.invalidate = function() {};
+Video.prototype.dispose = function() {};
+Video.prototype.clone = function() {};
