+++ src/anm/player.js
@@ -119,10 +119,11 @@ Player.DEFAULT_CONFIGURATION = { 'debug': false,
                                  'height': undefined,
                                  //'fps': undefined,
                                  'infiniteDuration': undefined, // undefined means 'auto'
-                                 'drawStill': undefined, // undefined means 'auto'
+                                 'drawStill': undefined, // undefined means 'auto',
                                  'audioEnabled': true,
                                  'audioGlobalVolume': 1.0,
                                  'imagesEnabled': true,
+                                 'videoEnabled': true,
                                  'shadowsEnabled': true,
                                  'handleEvents': undefined, // undefined means 'auto'
                                  'controlsEnabled': undefined, // undefined means 'auto'
